fingerprints dynaic decrement

+++ nebuia_plugin/src/main/java/com/distbit/nebuia_plugin/core/Finger.java
@@ -29,5 +29,5 @@ public class Finger {
                                  Bitmap indexOut, Bitmap middleOut, Bitmap ringOut, Bitmap littleOut
     );
 
-    public native int Quality(Bitmap bitmap);
+    public native float Quality(Bitmap bitmap);
 }
